+++ tests/integration/cli/test_project.py
@@ -41,7 +41,7 @@ class ProjectIntegrationScenarios(TestCase):
                 "-o",
                 result_dir,
                 "--",
-                "--save-images",
+                "--save-media",
                 "--reindex",
                 "1",
             )
@@ -72,7 +72,7 @@ class ProjectIntegrationScenarios(TestCase):
                 "-o",
                 result_dir,
                 "--",
-                "--save-images",
+                "--save-media",
             )
 
             self.assertTrue(osp.isdir(result_dir))
@@ -183,7 +183,7 @@ class ProjectIntegrationScenarios(TestCase):
                 result_dir,
                 "source-1",
                 "--",
-                "--save-images",
+                "--save-media",
             )
             parsed = Dataset.import_from(result_dir, "coco")
             compare_datasets(
@@ -227,7 +227,7 @@ class ProjectIntegrationScenarios(TestCase):
                 "-o",
                 result_dir,
                 "--",
-                "--save-images",
+                "--save-media",
             )
             parsed = Dataset.import_from(result_dir, "coco")
             compare_datasets(