+++ tests/unit/test_images.py
@@ -5,7 +5,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from datumaro.components.media import ByteImage, Image, RoIImage
+from datumaro.components.media import Image, RoIImage
 from datumaro.util.image import (
     encode_image,
     lazy_image,
@@ -191,74 +191,6 @@ class ImageTest(TestCase):
     def test_ext_detection_failure(self):
         image_bytes = b"\xff" * 10  # invalid image
         image = Image.from_bytes(data=image_bytes)
-        self.assertEqual(image.ext, None)
-
-
-class BytesImageTest(TestCase):
-    @staticmethod
-    def _gen_image_and_args_list(test_dir: str) -> Tuple[np.ndarray, bytes, List[Dict[str, Any]]]:
-        path = osp.join(test_dir, "path.png")
-        image = np.ones([2, 4, 3])
-        image_bytes = encode_image(image, "png")
-
-        return (
-            image,
-            image_bytes,
-            [
-                {"data": image_bytes},
-                {"data": lambda _: image_bytes},
-                {"data": lambda _: image_bytes, "ext": ".jpg"},
-                {"data": image_bytes, "path": path},
-                {"data": image_bytes, "path": path, "size": (2, 4)},
-                {"data": image_bytes, "path": path, "size": (2, 4)},
-                {"path": path},
-                {"path": path, "size": (2, 4)},
-            ],
-        )
-
-    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_lazy_image_shape(self):
-        data = encode_image(np.ones((5, 6, 3)), "png")
-
-        image_lazy = ByteImage(data=data, size=(2, 4))
-        image_eager = ByteImage(data=data)
-
-        self.assertEqual((2, 4), image_lazy.size)
-        self.assertEqual((5, 6), image_eager.size)
-
-    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_ctors(self):
-        with TestDir() as test_dir:
-            image, image_bytes, args_list = self._gen_image_and_args_list(test_dir)
-
-            for args in args_list:
-                with self.subTest(**args):
-                    img = ByteImage(**args)
-                    # pylint: disable=pointless-statement
-                    self.assertEqual("data" in args, img.has_data)
-                    if img.has_data:
-                        np.testing.assert_array_equal(img.data, image)
-                        self.assertEqual(img.get_bytes(), image_bytes)
-                    img.size
-                    if "size" in args:
-                        self.assertEqual(img.size, args["size"])
-                    if "ext" in args or "path" in args:
-                        self.assertEqual(img.ext, args.get("ext", ".png"))
-                    # pylint: enable=pointless-statement
-
-    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_ext_detection(self):
-        image_data = np.zeros((3, 4))
-
-        for ext in (".bmp", ".jpg", ".png"):
-            with self.subTest(ext=ext):
-                image = ByteImage(data=encode_image(image_data, ext))
-                self.assertEqual(image.ext, ext)
-
-    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_ext_detection_failure(self):
-        image_bytes = b"\xff" * 10  # invalid image
-        image = ByteImage(data=image_bytes)
         self.assertEqual(image.ext, None)