+++ CHANGELOG.md
@@ -31,12 +31,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/1120>)
 - Support a dictionary input in addition to a single image input for the model launcher to support Segment Anything Model
   (<https://github.com/openvinotoolkit/datumaro/pull/1133>)
+- Remove deprecates announced to be removed in 1.5.0
+  (<https://github.com/openvinotoolkit/datumaro/pull/1140>)
 
 ### Bug fixes
 - Fix bugs for Tile transform
   (<https://github.com/openvinotoolkit/datumaro/pull/1123>)
 - Disable Roboflow Tfrecord format when Tensorflow is not installed
   (<https://github.com/openvinotoolkit/datumaro/pull/1130>)
+- Raise VcsAlreadyExists error if vcs directory exists
+  (<https://github.com/openvinotoolkit/datumaro/pull/1138>)
 
 ## 27/07/2023 - Release 1.4.1
 ### Bug fixes